+++ app/src/main/java/com/easyfitness/graph/DateGraph.java
@@ -80,7 +80,7 @@ public class DateGraph {
         leftAxis.setTextColor(ColorTemplate.getHoloBlue());
         leftAxis.setDrawGridLines(true);
         leftAxis.setGranularityEnabled(true);
-        leftAxis.setGranularity((float) 0.5);
+        leftAxis.setGranularity(0.5f);
         leftAxis.setTextSize(12);
         leftAxis.resetAxisMinimum();
